removed SecretsManager from api

+++ src/pwnctl/pwnctl.dto/Mediator/MediatorError.cs
@@ -17,7 +17,7 @@ public sealed class MediatorError
 
     public int ToStatusCode()
     {
-        return int.Parse(Type.ToString().Substring(0, 3));
+        return int.Parse(((int)Type).ToString().Substring(0, 3));
     }
 
     public enum ErrorType
